test: Add BLI Status Change test
The commit adds a new test for BLI Status Change in the `statusChangeRequest.cy.js` file. This test covers the process of creating a test agreement, creating a BLI, submitting it for approval, and deleting the agreement.
