+++ frontend/cypress/e2e/budgetChangeRequest.cy.js
@@ -1,0 +1,136 @@
+/// <reference types="cypress" />
+
+import { BLI_STATUS } from "../../src/helpers/budgetLines.helpers";
+import { terminalLog, testLogin } from "./utils";
+
+const testAgreement = {
+    agreement_type: "CONTRACT",
+    agreement_reason: "NEW_REQ",
+    name: "E2E Test agreementWorkflow 1",
+    description: "Test Description",
+    project_id: 1,
+    product_service_code_id: 1,
+    procurement_shop_id: 1,
+    project_officer_id: 500,
+    team_members: [
+        {
+            id: 520
+        },
+        {
+            id: 504
+        }
+    ],
+    notes: "Test Notes"
+};
+
+const testBli = {
+    line_description: "SC1",
+    comments: "",
+    can_id: 1,
+    agreement_id: 11,
+    amount: 1_000_000,
+    status: BLI_STATUS.PLANNED,
+    date_needed: "2025-1-01",
+    proc_shop_fee_percentage: 0.005
+};
+
+beforeEach(() => {
+    testLogin("admin");
+    cy.visit(`/`);
+});
+
+afterEach(() => {
+    cy.injectAxe();
+    cy.checkA11y(null, null, terminalLog);
+});
+
+it("BLI Budget Change", () => {
+    expect(localStorage.getItem("access_token")).to.exist;
+
+    // create test agreement
+    const bearer_token = `Bearer ${window.localStorage.getItem("access_token")}`;
+    cy.request({
+        method: "POST",
+        url: "http://localhost:8080/api/v1/agreements/",
+        body: testAgreement,
+        headers: {
+            Authorization: bearer_token,
+            "Content-Type": "application/json",
+            Accept: "application/json"
+        }
+    })
+        .then((response) => {
+            expect(response.status).to.eq(201);
+            expect(response.body.id).to.exist;
+            const agreementId = response.body.id;
+            return agreementId;
+        })
+        // create BLI
+        .then((agreementId) => {
+            const bliData = { ...testBli, agreement_id: agreementId };
+            cy.request({
+                method: "POST",
+                url: "http://localhost:8080/api/v1/budget-line-items/",
+                body: bliData,
+                headers: {
+                    Authorization: bearer_token,
+                    Accept: "application/json"
+                }
+            }).then((response) => {
+                expect(response.status).to.eq(201);
+                expect(response.body.id).to.exist;
+                const bliId = response.body.id;
+                return { agreementId, bliId };
+            });
+        })
+        // submit for approval (via REST for now, maybe change to UI click through)
+        .then(({ agreementId, bliId }) => {
+            const payload = {
+                id: bliId,
+                amount: 2_000_000,
+                requestor_notes: "Please approve this budget change"
+            };
+            cy.request({
+                method: "PATCH",
+                url: `http://localhost:8080/api/v1/budget-line-items/${bliId}`,
+                body: payload,
+                headers: {
+                    Authorization: bearer_token,
+                    Accept: "application/json"
+                }
+            }).then((response) => {
+                expect(response.status).to.eq(202);
+                return { agreementId, bliId };
+            });
+        })
+        .then(({ agreementId, bliId }) => {
+            cy.visit("/agreements?filter=change-requests").wait(1000);
+            // see if there are any review cards review-card
+            cy.get("[data-cy='review-card']")
+                .should("exist")
+                .then(() => {
+                    cy.request({
+                        method: "DELETE",
+                        url: `http://localhost:8080/api/v1/budget-line-items/${bliId}`,
+                        headers: {
+                            Authorization: bearer_token,
+                            Accept: "application/json"
+                        }
+                    }).then((response) => {
+                        expect(response.status).to.eq(200);
+                    });
+                })
+                .then(() => {
+                    cy.request({
+                        method: "DELETE",
+                        url: `http://localhost:8080/api/v1/agreements/${agreementId}`,
+                        headers: {
+                            Authorization: bearer_token,
+                            Accept: "application/json"
+                        }
+                    }).then((response) => {
+                        expect(response.status).to.eq(200);
+                    });
+                });
+        });
+});
