+++ frontend/cypress/e2e/statusChangeRequest.cy.js
@@ -22,6 +22,7 @@ const testAgreement = {
     ],
     notes: "Test Notes"
 };
+
 const testBli = {
     line_description: "SC1",
     comments: "",
@@ -29,16 +30,6 @@ const testBli = {
     agreement_id: 11,
     amount: 1_000_000,
     status: BLI_STATUS.DRAFT,
-    date_needed: "2025-1-01",
-    proc_shop_fee_percentage: 0.005
-};
-const testBLI2 = {
-    line_description: "SC2",
-    comments: "",
-    can_id: 1,
-    agreement_id: 11,
-    amount: 2_000_000,
-    status: BLI_STATUS.PLANNED,
     date_needed: "2025-1-01",
     proc_shop_fee_percentage: 0.005
 };
@@ -97,85 +88,6 @@ it("BLI Status Change", () => {
             const payload = {
                 id: bliId,
                 status: BLI_STATUS.PLANNED,
-                requestor_notes: "Please approve this BLI"
-            };
-            cy.request({
-                method: "PATCH",
-                url: `http://localhost:8080/api/v1/budget-line-items/${bliId}`,
-                body: payload,
-                headers: {
-                    Authorization: bearer_token,
-                    Accept: "application/json"
-                }
-            }).then((response) => {
-                expect(response.status).to.eq(202);
-                return { agreementId, bliId };
-            });
-        })
-        .then(({ agreementId, bliId }) => {
-            cy.visit("/agreements?filter=change-requests").wait(1000);
-            // see if there are any review cards review-card
-            cy.get("[data-cy='review-card']")
-                .should("exist")
-                .then(() => {
-                    cy.request({
-                        method: "DELETE",
-                        url: `http://localhost:8080/api/v1/agreements/${agreementId}`,
-                        headers: {
-                            Authorization: bearer_token,
-                            Accept: "application/json"
-                        }
-                    }).then((response) => {
-                        expect(response.status).to.eq(200);
-                    });
-                });
-        });
-});
-
-it("BLI Budget Change", () => {
-    expect(localStorage.getItem("access_token")).to.exist;
-
-    // create test agreement
-    const bearer_token = `Bearer ${window.localStorage.getItem("access_token")}`;
-    cy.request({
-        method: "POST",
-        url: "http://localhost:8080/api/v1/agreements/",
-        body: testAgreement,
-        headers: {
-            Authorization: bearer_token,
-            "Content-Type": "application/json",
-            Accept: "application/json"
-        }
-    })
-        .then((response) => {
-            expect(response.status).to.eq(201);
-            expect(response.body.id).to.exist;
-            const agreementId = response.body.id;
-            return agreementId;
-        })
-        // create BLI
-        .then((agreementId) => {
-            const bliData = { ...testBLI2, agreement_id: agreementId };
-            cy.request({
-                method: "POST",
-                url: "http://localhost:8080/api/v1/budget-line-items/",
-                body: bliData,
-                headers: {
-                    Authorization: bearer_token,
-                    Accept: "application/json"
-                }
-            }).then((response) => {
-                expect(response.status).to.eq(201);
-                expect(response.body.id).to.exist;
-                const bliId = response.body.id;
-                return { agreementId, bliId };
-            });
-        })
-        // submit for approval (via REST for now, maybe change to UI click through)
-        .then(({ agreementId, bliId }) => {
-            const payload = {
-                id: bliId,
-                amount: 3_000_000,
                 requestor_notes: "Please approve this BLI"
             };
             cy.request({
